backend/requirements.txt
aiohttp==3.8.4
asyncpg==0.27.0
python-dotenv==0.19.2
spacy==3.5.0
torch==1.13.1
transformers==4.26.1
textblob==0.17.1
scikit-learn==1.2.1
pydantic==1.10.7